pyproject.toml
[project]
name = "builder-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langgraph>=0.5.0",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
